+++ projet_web-main/SCRIPTS/suppr.php
@@ -8,18 +8,34 @@ $DB_NAME = getenv('DB_NAME');
 $DB_USER = getenv('DB_USER');
 $DB_PASS = getenv('DB_PASS');
 
-$seance=$_POST['idSeance'];
+session_start();
+if(isset($_SESSION['idUser'])){
+    if($_SESSION['idUser'] == 0){
 
-if(!empty($seance)){
-    $cnx= Connexion($DB_HOST,$DB_NAME,$DB_USER,$DB_PASS);
-    
-    $req = "DELETE FROM seance WHERE idSeance= '$seance'";
-    $result=requeteSelect($cnx, $req);
-    
-    header("Location: ../PAGES/Visu_tab_R.php?succes=Séance supprimée avec succès");
-}
-else {
-    header("Location: ../PAGES/suppr_seance.php?error=Veuillez sélectionner une séance");
-}
+        $idUser = $_SESSION['idUser'];
 
-?>
+        if(isset($_POST['idSeance'])){
+            $seance=$_POST['idSeance'];
+        }
+        
+        if(!empty($seance)){
+            $cnx= Connexion($DB_HOST,$DB_NAME,$DB_USER,$DB_PASS);
+            
+            $req = "DELETE FROM seance WHERE idSeance= '$seance'";
+            $result=requeteSelect($cnx, $req);
+            
+            header("Location: ../PAGES/Visu_tab_R.php?succes=Séance supprimée avec succès");
+        }
+        else {
+            header("Location: ../PAGES/suppr_seance.php?error=Veuillez sélectionner une séance");
+        }
+    }
+    
+    else{
+        header("Location: ../PAGES/Visu_tab_IC.php");
+    }
+}
+else{
+    header("Location: ../index.php");
+}
+        ?>
